Finished Upgrade Interface
Interface now functions correctly.
Avoided a near catastrophic crash
Intermediary files replaced.
Discovered an error in Unreal Engine 5 where
updating a c++ created component  sets the component to null in runtime which crashes the game and editor. The only work around I've found at the moment is to comment out the c++ implementation. Save and compile it then when the component is fully deattached uncomment out the implementation so it refreshes.